Refs #23739: Move tutorial, fix imports, spelling
