+++ docs/index.rst
@@ -100,6 +100,7 @@ To make such contributions, please refer to the
    /rst/microros_documentation/index.rst
    /rst/enhancements/intro.rst
    /rst/hri/intro.rst
+   /rst/vulcanai/vulcanai_introduction.rst
 
 .. toctree::
    :caption: Tutorials
